(MAINT) Add windows script
Prior to this commit the project included a script for running
the beaker container on *nix systems using bash - this commit
adds an equivalent script in powershell for windowsy systems.

It also ensures that the bundle_beaker script will always have
LF line endings to prevent problems when running on windows.

Finally, it modifies the script and dockerfile to use bash and
to drop the container into the module folder to start with.